requirements.txt
mutagen
openai-whisper
